improve qingcloud errors
